#23: implement user controller, service, and repository with TypeScript
